chore(TS): Updated types structure
