+++ src/navigation/settings-items.ts
@@ -1,7 +1,7 @@
 import { PROFILE, USERS, ACCOUNTS, CLOUD_FUNCTIONS } from '@/constants/router/routes-names';
 import { SUPER_ADMIN } from '@/constants/user/roles';
 import store from '@/store';
-import { Link } from '@/types/link';
+import { Link } from '@/types';
 
 const isSuperAdminRule = () => store.getters.user.studioRoles === SUPER_ADMIN.roleCode;
