--- src/types/header-item.ts
@@ -1,5 +1,0 @@
-export interface HeaderItem {
-	text: string;
-	sortable: boolean;
-	value: string;
-}
