--- src/types/link.ts
@@ -1,8 +1,0 @@
-import { Location } from 'vue-router';
-
-export interface Link {
-	icon: string;
-	title: string;
-	link: Location;
-	displayRule?(): boolean;
-}
