--- src/types/role.ts
@@ -1,4 +1,0 @@
-export interface Role {
-	roleName: string;
-	roleCode: 0 | 1 | 2 | 3 | 4 | 5;
-}
